requirements.txt
asgiref==3.7.2
boto3==1.29.6
Django==4.2.7
numpy==1.26.2
opencv-python==4.8.1.78
Pillow==10.1.0
psycopg2
pytest==7.4.3
sqlparse==0.4.4
django-widget-tweaks